# docs/swagger.yaml
basePath: /api/v0
definitions:
...
  types.EpisodeImages:
    properties:
      medium:
        type: string
      original:
        type: string
    type: object
  types.EpisodeLink:
    properties:
      href:
        type: string
    type: object
  types.EpisodeLinks:
    properties:
      self:
        $ref: '#/definitions/types.EpisodeLink'
    type: object
...
  types.MovieTorrent:
    properties:
      hash:
        type: string
      lang:
        type: string
      magnet:
        type: string
      peers:
        type: string
      provider:
        type: string
      quality:
        type: string
      seeds:
        type: string
      size:
        type: string
      title:
        type: string
      torrent:
        type: string
    type: object
...
  types.ShowTorrent:
    properties:
      episode:
        type: string
      hash:
        type: string
      lang:
        type: string
      magnet:
        type: string
      peers:
        type: string
      provider:
        type: string
      quality:
        type: string
      season:
        type: string
      seeds:
        type: string
      size:
        type: string
      title:
        type: string
      torrent:
        type: string
    type: object
...
  types.TvMazeEpisode:
    properties:
      _links:
        $ref: '#/definitions/types.EpisodeLinks'
      airdate:
        type: string
      airstamp:
        type: string
      airtime:
        type: string
      id:
        type: integer
      image:
        $ref: '#/definitions/types.EpisodeImages'
      inumberd:
        type: integer
      name:
        type: string
      runtime:
        type: integer
      season:
        type: integer
      summary:
        type: string
      type:
        type: string
      url:
        type: string
    type: object
...
  v0.MessageResponse:
    properties:
      message:
        type: string
      success:
        type: boolean
    type: object
  v0.MovieMagnetLinksResponse:
    properties:
      results:
        items:
          $ref: '#/definitions/types.MovieTorrent'
        type: array
      success:
        type: boolean
    type: object
  v0.ShowEpisodesResponse:
    properties:
      results:
        items:
          $ref: '#/definitions/types.TvMazeEpisode'
        type: array
      success:
        type: boolean
    type: object
  v0.ShowMagnetLinksResponse:
    properties:
      results:
        items:
          $ref: '#/definitions/types.ShowTorrent'
        type: array
      success:
        type: boolean
    type: object
...
host: localhost:9000
info:
  contact: {}
  title: Raven Torrent API
  version: 0.6.0
paths:
...
  /getmoviemagnet/imdb/{imdb}/providers/{providers}:
    get:
      parameters:
      - description: ' '
        in: path
        name: imdb
        required: true
        type: string
      - description: ' '
        in: path
        name: providers
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v0.MovieMagnetLinksResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v0.MessageResponse'
      summary: Get movie torrents by IMDB id
      tags:
      - Torrent search
  /getmoviemagnet/imdb/{imdb}/query/{query}/providers/{providers}:
    get:
      parameters:
      - description: ' '
        in: path
        name: imdb
        required: true
        type: string
      - description: ' '
        in: path
        name: query
        required: true
        type: string
      - description: ' '
        in: path
        name: providers
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v0.MovieMagnetLinksResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v0.MessageResponse'
      summary: Get movie torrents by IMDB id and query string
      tags:
      - Torrent search
  /getmoviemagnet/query/{query}/providers/{providers}:
    get:
      parameters:
      - description: ' '
        in: path
        name: query
        required: true
        type: string
      - description: ' '
        in: path
        name: providers
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v0.MovieMagnetLinksResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v0.MessageResponse'
      summary: Get movie torrents by query string
      tags:
      - Torrent search
  /getshowmagnet/imdb/{imdb}/query/{query}/season/{season}/episode/{episode}/providers/{providers}:
    get:
      parameters:
      - description: ' '
        in: path
        name: imdb
        required: true
        type: string
      - description: ' '
        in: path
        name: query
        required: true
        type: string
      - description: ' '
        in: path
        name: season
        required: true
        type: integer
      - description: ' '
        in: path
        name: episode
        required: true
        type: integer
      - description: ' '
        in: path
        name: providers
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v0.ShowMagnetLinksResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v0.MessageResponse'
      summary: Get show torrents by IMDB id and query string
      tags:
      - Torrent search
  /getshowmagnet/imdb/{imdb}/season/{season}/episode/{episode}/providers/{providers}:
    get:
      parameters:
      - description: ' '
        in: path
        name: imdb
        required: true
        type: string
      - description: ' '
        in: path
        name: season
        required: true
        type: integer
      - description: ' '
        in: path
        name: episode
        required: true
        type: integer
      - description: ' '
        in: path
        name: providers
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v0.ShowMagnetLinksResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v0.MessageResponse'
      summary: Get show torrents by IMDB id
      tags:
      - Torrent search
  /getshowmagnet/query/{query}/season/{season}/episode/{episode}/providers/{providers}:
    get:
      parameters:
      - description: ' '
        in: path
        name: query
        required: true
        type: string
      - description: ' '
        in: path
        name: season
        required: true
        type: integer
      - description: ' '
        in: path
        name: episode
        required: true
        type: integer
      - description: ' '
        in: path
        name: providers
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v0.ShowMagnetLinksResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v0.MessageResponse'
      summary: Get show torrents by query string
      tags:
      - Torrent search
...
  /tvmazeepisodes/imdb/{imdb}:
    get:
      parameters:
      - description: ' '
        in: path
        name: imdb
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v0.ShowEpisodesResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v0.MessageResponse'
      summary: Get show episodes by IMDB id
      tags:
      - Media search
  /tvmazeepisodes/tvdb/{tvdb}:
    get:
      parameters:
      - description: ' '
        in: path
        name: tvdb
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v0.ShowEpisodesResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v0.MessageResponse'
      summary: Get show episodes by TVDB id
      tags:
      - Media search
  /tvmazeepisodes/tvdb/{tvdb}/imdb/{imdb}:
    get:
      parameters:
      - description: ' '
        in: path
        name: tvdb
        required: true
        type: string
      - description: ' '
        in: path
        name: imdb
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v0.ShowEpisodesResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v0.MessageResponse'
      summary: Get show episodes by TVDB id and IMDB id
      tags:
      - Media search
swagger: "2.0"